% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para_gen.R
\name{para_gen}
\alias{para_gen}
\title{Generate simulation parameters}
\usage{
para_gen(model, seed = 999)
}
\arguments{
\item{model}{model number}

\item{seed}{random seed}
}
\value{
List of parameters
}
\description{
Generate simulation parameters
}
\examples{
para_gen(model=1)
}
